+++ packages/datadog_flutter_plugin/lib/datadog_flutter_plugin.dart
@@ -61,7 +61,6 @@ class DatadogSdk {
   void _setFirstPartyHosts(List<String> value) {
     _firstPartyHosts = value;
     if (value.isNotEmpty) {
-      // pattern = "^(.*\\.)*tracedHost1$|tracedHost2$|...$"
       var hosts = value.map((e) => '${RegExp.escape(e)}\$').join('|');
       _firstPartyRegex = RegExp('^(.*\\.)*$hosts');
     } else {